+++ veterans/components/homePage/mapSection/MapSection.tsx
@@ -30,10 +30,7 @@ export default function MapSection() {
   return (
     <section className={st.container}>
       <div className={st['wrapper-filter']}>
-        <RegionsList
-          selectedRegion={selectedRegion}
-          setSelectedRegion={setSelectedRegion}
-        />
+        <RegionsList setCurrentRegion={setSelectedRegion} />
         <InitiativesFilter
           selectedCheckboxes={selectedCheckboxes}
           setSelectedCheckboxes={setSelectedCheckboxes}