+++ veterans/components/homePage/mapSection/regionsFilter/types.ts
@@ -1,0 +1,8 @@
+export interface Region {
+  name: string;
+  numOfInitiatives?: number;
+}
+
+export interface RegionsListProps {
+  setCurrentRegion?: (region: string) => void;
+}